add areafilter POM and use it
