+++ frontend/fingertips-frontend/playwright/page-objects/pages/resultsPage.ts
@@ -1,14 +1,12 @@
 import { SearchParams } from '@/lib/searchStateManager';
-import BasePage from '../basePage';
 import { expect } from '../pageFactory';
 import {
-  AreaMode,
   IndicatorMode,
   returnIndicatorIDsByIndicatorMode,
-  SearchMode,
 } from '@/playwright/testHelpers';
+import AreaFilter from '../components/areaFilter';
 
-export default class ResultsPage extends BasePage {
+export default class ResultsPage extends AreaFilter {
   readonly resultsText = 'Search results for';
   readonly backLink = 'search-results-back-link';
   readonly searchResult = 'search-result';
@@ -20,14 +18,6 @@ export default class ResultsPage extends BasePage {
   readonly indicatorSearchError = `indicator-search-form-error`;
   readonly indicatorSearchButton = `indicator-search-form-submit`;
   readonly indicatorCheckboxContainer = 'indicator-selection-form';
-  readonly areaFilterContainer = 'area-filter-container';
-  readonly areaTypeSelector = 'area-type-selector-container';
-  readonly groupTypeSelector = 'group-type-selector-container';
-  readonly selectedAreasContainer = 'selected-areas-panel';
-  readonly selectAreasContainer = 'select-areas-filter-panel';
-  readonly pillContainer = 'pill-container';
-  readonly filterName = 'filter-name';
-  readonly removeIcon = 'x-icon';
 
   async navigateToResults(
     searchIndicator: string,
@@ -42,48 +32,6 @@ export default class ResultsPage extends BasePage {
     await this.page.goto(
       `results?${SearchParams.SearchedIndicator}=${searchIndicator}${asQuery}`
     );
-  }
-
-  areaFilterPills() {
-    return this.page
-      .getByTestId(this.areaFilterContainer)
-      .getByTestId(this.pillContainer);
-  }
-
-  async areaFilterPillsText() {
-    const pillFilterNames = await this.areaFilterPills()
-      .getByTestId(this.filterName)
-      .all();
-
-    return Promise.all(pillFilterNames.map(async (l) => await l.textContent()));
-  }
-
-  async closeAreaFilterPill(index: number) {
-    const pills = await this.areaFilterPills()
-      .getByTestId(this.removeIcon)
-      .all();
-
-    await pills[index].click();
-  }
-
-  areaFilterCombobox() {
-    return this.page
-      .getByTestId(this.areaFilterContainer)
-      .getByTestId(this.areaTypeSelector)
-      .locator('select');
-  }
-
-  areaFilterOptions() {
-    return this.page
-      .getByTestId(this.areaFilterContainer)
-      .getByTestId(this.areaTypeSelector)
-      .getByRole('option');
-  }
-
-  async areaFilterOptionsText() {
-    const options = await this.areaFilterOptions().all();
-
-    return Promise.all(options.map((l) => l.textContent()));
   }
 
   async checkSearchResultsTitle(searchTerm: string) {
@@ -148,69 +96,6 @@ export default class ResultsPage extends BasePage {
 
       await expect(checkbox).toBeChecked();
       await this.waitForURLToContain(indicatorID);
-    }
-  }
-
-  /**
-   * Selects the required area filters based on area mode if the search mode is ONLY_SUBJECT
-   *
-   * @param areaMode - area mode from the Enum AreaMode - used to decide which area filters to select
-   * @param searchTerm - search term to be used in the URL check
-   */
-  async selectAreasFiltersIfRequired(
-    searchMode: SearchMode,
-    areaMode: AreaMode,
-    searchTerm: string
-  ) {
-    // For area type filter currently defaulting to using regions (except for England area mode) - this will be refactored in DHSCFT-416
-    const defaultAreaTypeFilter = 'regions';
-
-    await this.waitForURLToContain(searchTerm);
-
-    // only do the following for SearchMode.ONLY_SUBJECT as SearchMode.ONLY_AREA/BOTH_SUBJECT_AND_AREA already have area filters selected
-    if (searchMode === SearchMode.ONLY_SUBJECT) {
-      await this.page
-        .getByTestId(this.areaTypeSelector)
-        .selectOption(defaultAreaTypeFilter);
-
-      await this.waitForURLToContain(defaultAreaTypeFilter);
-
-      // For group type filter currently defaults to using England due to picking regions for area type above - this will be refactored in DHSCFT-416
-
-      // Select appropriate number of checkboxes based on area mode
-      const areaCheckboxList = this.page
-        .getByTestId(this.areaFilterContainer)
-        .getByRole('checkbox');
-      const checkboxCountMap = {
-        [AreaMode.ONE_AREA]: 1,
-        [AreaMode.TWO_PLUS_AREAS]: 2,
-        [AreaMode.ALL_AREAS_IN_A_GROUP]: (await areaCheckboxList.count()) - 1, // as first checkbox is 'All'
-        [AreaMode.ENGLAND_AREA]: 0, // for england we do not want to select any checkboxes
-      };
-      const checkboxCount = checkboxCountMap[areaMode];
-      for (let i = 0; i < checkboxCount; i++) {
-        await areaCheckboxList.nth(i + 1).check(); // as first checkbox is 'All'
-        await this.page.waitForLoadState();
-        if (i === 0 && areaMode !== AreaMode.ENGLAND_AREA) {
-          await this.waitForURLToContain(defaultAreaTypeFilter);
-        }
-      }
-      await expect(
-        this.page.getByTestId(this.areaFilterContainer)
-      ).toContainText(`Selected areas (${String(checkboxCount)})`);
-
-      // England area mode
-      if (AreaMode.ENGLAND_AREA === areaMode) {
-        await this.page
-          .getByTestId(this.areaTypeSelector)
-          .selectOption('England');
-        await this.page
-          .getByTestId(this.groupTypeSelector)
-          .selectOption('England');
-        await this.waitForURLToContain('England');
-      }
-
-      await this.waitForURLToContain(searchTerm);
     }
   }
